commit 2

+++ ACT23.C
@@ -2,14 +2,15 @@
 
 int main (void)
 {
-    int cat1;
-    int cat2;
+    float cat1;
+    float cat2;
     printf("ingrese la medida de un cateto de su triangulo\n");
-    scanf("%d",&cat1);
+    scanf("%f",&cat1);
     printf("ingrese la medida del otro cateto\n");
-    scanf("%d",&cat2);
-    int area = (cat1 * cat2)/2;
-    printf("el area de su triangulo es de %d\n",area);
+    scanf("%f",&cat2);
+    float area = (cat1 * cat2)/2;
+    printf("el area de su triangulo es de %f\n",area);
 
     return 0;
+
 }
